Refactor: Move RootScreen to MainActivity
The `RootScreen` composable has been moved into the `MainActivity` to simplify the UI structure.

The `BadHabitsScreen` composable, previously part of `RootScreen.kt`, has been moved to its own file `BadHabitsScreen.kt` within the `badhabits.ui` package.
