+++ app/src/main/java/com/kawunus/habityou/ui/common/navigation/ui/NavigationHost.kt
@@ -10,11 +10,11 @@ import com.kawunus.habityou.ui.common.navigation.model.BottomNavItem
 import com.kawunus.habityou.ui.common.navigation.model.NavigationConstants.EDIT_NOTE_ROUTE
 import com.kawunus.habityou.ui.common.navigation.model.NavigationConstants.NEW_NOTE_ROUTE
 import com.kawunus.habityou.ui.common.navigation.model.NavigationConstants.NEW_USEFUL_HABIT_ROUTE
+import com.kawunus.habityou.ui.screens.badhabits.ui.BadHabitsScreen
 import com.kawunus.habityou.ui.screens.diary.ui.DiaryScreen
 import com.kawunus.habityou.ui.screens.editnote.ui.EditNoteScreen
 import com.kawunus.habityou.ui.screens.newnote.ui.NewNoteScreen
 import com.kawunus.habityou.ui.screens.newusefulhabit.ui.NewUsefulHabitScreen
-import com.kawunus.habityou.ui.screens.root.ui.BadHabitsScreen
 import com.kawunus.habityou.ui.screens.usefulhabits.ui.UsefulHabitsScreen
